+++ config/test.exs
@@ -1,1 +1,4 @@
 use Mix.Config
+
+config :ex_unit,
+  capture_log: true