+++ lib/json_api_client.ex
@@ -4,11 +4,13 @@ defmodule JsonApiClient do
   the JSON API spec described at http://jsonapi.org
   """
 
-  @client_name Application.get_env(:json_api_client, :client_name)
   @timeout Application.get_env(:json_api_client, :timeout, 500)
   @version Mix.Project.config[:version]
+  @package_name JsonApiClient.Mixfile.project[:app]
 
-  alias __MODULE__.{Request, RequestError, Response, Parser}
+  alias __MODULE__.Request
+  alias __MODULE__.Middleware.Runner
+  alias Mix.Project
 
   @doc "Execute a JSON API Request using HTTP GET"
   def fetch(req), do: req |> Request.method(:get) |> execute
@@ -34,7 +36,7 @@ defmodule JsonApiClient do
   Execute a JSON API Request
 
   Takes a JsonApiClient.Request and preforms the described request.
-  
+
   Returns either a tuple with `:ok` and a `JsonApiClient.Response` struct (or
   nil) or `:error` and a `JsonApiClient.RequestError` struct depending on the
   http response code and whether the server response was valid according to the
@@ -52,22 +54,6 @@ defmodule JsonApiClient do
 
   """
   def execute(req) do
-    with {:ok, response} <- do_request(req),
-         {:ok, parsed}   <- parse_response(response)
-    do
-      {:ok, parsed}
-    end
-  end
-
-  @doc "Error raising version of `execute/1`"
-  def execute!(req) do
-    case execute(req) do
-      {:ok, response} -> response
-      {:error, error} -> raise error
-    end
-  end
-
-  defp do_request(req) do
     url          = Request.get_url(req)
     query_params = Request.get_query_params(req)
     headers      = default_headers()
@@ -79,36 +65,22 @@ defmodule JsonApiClient do
                    |> Enum.into([])
     body = Request.get_body(req)
 
-    case HTTPoison.request(req.method, url, body, headers, http_options) do
-      {:ok, _} = result -> result
-      {:error, error} ->
-        {:error, %RequestError{
-          original_error: error,
-          message: "Error completing HTTP request: #{error.reason}",
-        }}
-    end
+    Runner.run %{
+      method: req.method,
+      url: url, body: body,
+      headers: headers,
+      http_options: http_options,
+      service_name: req.service_name
+    }
   end
 
-  defp parse_response(response) do
-    with {:ok, doc} <- parse_document(response.body)
-    do
-      {:ok, %Response{
-        status: response.status_code,
-        doc: doc,
-        headers: response.headers,
-      }}
-    else
-      {:error, error} ->
-        {:error, %RequestError{
-          message: "Error Parsing JSON API Document",
-          original_error: error,
-          status: response.status_code,
-        }}
+  @doc "Error raising version of `execute/1`"
+  def execute!(req) do
+    case execute(req) do
+      {:ok, response} -> response
+      {:error, error} -> raise error
     end
   end
-
-  defp parse_document(""), do: {:ok, nil}
-  defp parse_document(json), do: Parser.parse(json)
 
   defp default_options do
     %{
@@ -126,7 +98,13 @@ defmodule JsonApiClient do
   end
 
   defp user_agent do
-    "ExApiClient/" <> @version <> "/" <> @client_name
+    [@package_name, @version, user_agent_suffix()]
+      |> Enum.reject(&is_nil/1)
+      |> Enum.join("/")
+  end
+
+  defp user_agent_suffix do
+    Application.get_env(:json_api_client, :user_agent_suffix, Project.config[:app])
   end
 
   defp timeout do