+++ lib/json_api_client/response.ex
@@ -6,7 +6,13 @@ defmodule JsonApiClient.Response do
   * status - The HTTP Status code from the response
   * headers - The HTTP Headers as a list of tuples `[{header, value}, ...]`
   * doc - A `JsonApiClient.Document` if one was present in the response or nil
+  * attributes - Custom attributes.
   """
 
-  defstruct [:status, :headers, :doc]
+  defstruct(
+    status: nil,
+    headers: nil,
+    doc: nil,
+    attributes: %{}
+  )
 end